+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/google_maps_flutter_ios/Sources/google_maps_flutter_ios/include/google_maps_flutter_ios/FGMClusterManagersController.h
@@ -6,6 +6,7 @@
 @import GoogleMaps;
 @import GoogleMapsUtils;
 
+#import "FGMMapEventDelegate.h"
 #import "google_maps_flutter_pigeon_messages.g.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -15,10 +16,10 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// Initializes cluster manager controller.
 ///
-/// @param callbackHandler A callback handler.
+/// @param eventDelegate A delegate that will receive events from the cluster managers.
 /// @param mapView A map view that will be used to display clustered markers.
 - (instancetype)initWithMapView:(GMSMapView *)mapView
-                callbackHandler:(FGMMapsCallbackApi *)callbackHandler;
+                  eventDelegate:(NSObject<FGMMapEventDelegate> *)eventDelegate;
 
 /// Creates cluster managers and initializes them.
 ///