+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/google_maps_flutter_ios/Sources/google_maps_flutter_ios/include/google_maps_flutter_ios/GoogleMapMarkerController_Test.h
@@ -16,7 +16,7 @@
 + (void)updateMarker:(GMSMarker *)marker
            fromPlatformMarker:(FGMPlatformMarker *)platformMarker
                   withMapView:(GMSMapView *)mapView
-                    registrar:(NSObject<FlutterPluginRegistrar> *)registrar
+                assetProvider:(NSObject<FGMAssetProvider> *)assetProvider
                   screenScale:(CGFloat)screenScale
     usingOpacityForVisibility:(BOOL)useOpacityForVisibility;