+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/include/video_player_avfoundation/FVPDisplayLink.h
@@ -2,13 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import <Foundation/Foundation.h>
+@import Foundation;
 
-#if TARGET_OS_OSX
-#import <FlutterMacOS/FlutterMacOS.h>
-#else
-#import <Flutter/Flutter.h>
-#endif
+#import "FVPViewProvider.h"
 
 // A cross-platform display link abstraction.
 @protocol FVPDisplayLink <NSObject>
@@ -31,8 +27,8 @@ API_AVAILABLE(ios(4.0), macos(14.0))
 ///
 /// The display link starts paused, so must be started, by setting 'running' to YES, before the
 /// callback will fire.
-- (instancetype)initWithRegistrar:(id<FlutterPluginRegistrar>)registrar
-                         callback:(void (^)(void))callback NS_DESIGNATED_INITIALIZER;
+- (instancetype)initWithViewProvider:(NSObject<FVPViewProvider> *)viewProvider
+                            callback:(void (^)(void))callback NS_DESIGNATED_INITIALIZER;
 
 - (instancetype)init NS_UNAVAILABLE;
 
@@ -46,8 +42,8 @@ API_AVAILABLE(ios(4.0), macos(14.0))
 ///
 /// The display link starts paused, so must be started, by setting 'running' to YES, before the
 /// callback will fire.
-- (instancetype)initWithRegistrar:(id<FlutterPluginRegistrar>)registrar
-                         callback:(void (^)(void))callback NS_DESIGNATED_INITIALIZER;
+- (instancetype)initWithViewProvider:(NSObject<FVPViewProvider> *)viewProvider
+                            callback:(void (^)(void))callback NS_DESIGNATED_INITIALIZER;
 
 - (instancetype)init NS_UNAVAILABLE;